payment details completed

+++ Models/CardDetails.cs
@@ -7,7 +7,7 @@ namespace Payment.Models
 {
     public class CardDetails
     {
-        public int Card_No { get; set; }
+        public long Card_No { get; set; }
         public int Card_Limit { get; set; }
 
     }
